backend-py/pyproject.toml
[tool.poetry]
name = "rdweb-moderno-backend"
version = "1.0.0"
description = "API REST para RDWeb-Moderno — reemplazo de RD Web Access (Python/FastAPI)"
authors = ["RDS Custom"]
readme = "README.md"
packages = [{ include = "app" }]

[tool.poetry.dependencies]
python = ">=3.11,<3.15"
fastapi = "^0.115"
uvicorn = { version = "^0.34", extras = ["standard"] }
python-dotenv = "^1.0"
python-jose = { version = "^3.3", extras = ["cryptography"] }
ldap3 = "^2.9"
pydantic = "^2.10"
pydantic-settings = "^2.7"

[tool.poetry.group.dev.dependencies]
pyinstaller = "^6.11"
ruff = "^0.8"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 120
target-version = "py311"
